feat: move parsing of different file types into different features
