+++ crates/flatzinc-serde/Cargo.toml
@@ -15,10 +15,15 @@ exclude = ["/corpus"]
 
 edition = "2021"
 
+[features]
+default = ["serde"]
+serde = ["dep:serde"]
+fzn = ["dep:winnow"]
+
 [dependencies]
-serde = { version = "1.0", features = ["derive"] }
+serde = { version = "1.0", features = ["derive"], optional = true }
 rangelist = { path = "../rangelist", version = "0.4.0" }
-winnow = { version = "0.7.15" }
+winnow = { version = "0.7.15", optional = true }
 
 [dev-dependencies]
 serde_json = "1.0"
