+++ crates/flatzinc-serde/src/lib.rs
@@ -13,6 +13,13 @@
 //! in this crate to parse the FlatZinc JSON files produced by the MiniZinc
 //! compiler.
 //!
+//! # Feature Flags
+//!
+//! - `serde` (default): enables JSON serialization and deserialization support
+//!   via the [`serde`](https://serde.rs) crate.
+//! - `fzn`: enables parsing of the original `.fzn` text format via
+//!   [`winnow`](https://crates.io/crates/winnow).
+//!
 //! # Getting Started
 //!
 //! Install `flatzinc-serde` and `serde_json` for your package:
@@ -25,6 +32,7 @@
 //! deserialize a FlatZinc JSON file as follows:
 //!
 //! ```
+//! # #[cfg(feature = "serde")] {
 //! # use flatzinc_serde::FlatZinc;
 //! # use std::{fs::File, io::BufReader, path::Path};
 //! # let path = Path::new("./corpus/json/documentation_example.fzn.json");
@@ -32,16 +40,19 @@
 //! let rdr = BufReader::new(File::open(path).unwrap());
 //! let fzn: FlatZinc = serde_json::from_reader(rdr).unwrap();
 //! // ... process FlatZinc ...
+//! # }
 //! ```
 //!
 //! If, however, you want to serialize a FlatZinc format you could follow the
 //! following fragment:
 //!
 //! ```
+//! # #[cfg(feature = "serde")] {
 //! # use flatzinc_serde::FlatZinc;
 //! let fzn = FlatZinc::<String>::default();
 //! // ... create  solver constraint model ...
 //! let json_str = serde_json::to_string(&fzn).unwrap();
+//! # }
 //! ```
 //! Note that `serde_json::to_writer`, using a buffered file writer, would be
 //! preferred when writing larger FlatZinc files.
@@ -77,27 +88,18 @@
 //! (e.g., `../../../bin/fzn-my-solver model.fzn.json`).
 
 #![warn(missing_docs)]
-#![warn(unused_crate_dependencies, unused_extern_crates)]
 #![warn(variant_size_differences)]
 
+#[cfg(feature = "fzn")]
 pub mod fzn;
+#[cfg(feature = "serde")]
 mod serde;
 
 use std::{collections::BTreeMap, fmt::Display};
 
-use ::serde::{de::Error as SerdeError, Deserialize, Deserializer, Serialize, Serializer};
+#[cfg(feature = "serde")]
+use ::serde::{Deserialize, Serialize};
 pub use rangelist::RangeList;
-
-use crate::serde::{
-	deserialize_encapsulated_set, deserialize_encapsulated_string, deserialize_key_value_object,
-	serialize_encapsulate_set, serialize_encapsulate_string, serialize_key_value_object, BaseType,
-	VariableDomain,
-};
-
-/// Helper function to help skip in serialization
-fn is_false(b: &bool) -> bool {
-	!(*b)
-}
 
 /// Additional information provided in a standardized format for declarations,
 /// constraints, or solve objectives
@@ -108,8 +110,9 @@ fn is_false(b: &bool) -> bool {
 /// Note that annotations are generally defined either in the MiniZinc standard
 /// library or in a solver's redefinition library. Solvers are encouraged to
 /// rewrite annotations in their redefinitions library when required.
-#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
-#[serde(untagged)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(untagged))]
+#[derive(Clone, PartialEq, Debug)]
 pub enum Annotation<Identifier = String> {
 	/// Atom annotation (i.e., a single `Identifier`)
 	Atom(Identifier),
@@ -128,8 +131,9 @@ impl<Identifier: Display> Display for Annotation<Identifier> {
 }
 
 /// The argument type associated with [`AnnotationCall`]
-#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
-#[serde(untagged)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(untagged))]
+#[derive(Clone, PartialEq, Debug)]
 pub enum AnnotationArgument<Identifier = String> {
 	/// Sequence of [`Literal`]s
 	Array(Vec<AnnotationLiteral<Identifier>>),
@@ -158,8 +162,9 @@ impl<Idenfier: Display> Display for AnnotationArgument<Idenfier> {
 }
 
 /// An object depicting an annotation in the form of a call
-#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
-#[serde(rename = "annotation_call")]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(rename = "annotation_call"))]
+#[derive(Clone, PartialEq, Debug)]
 pub struct AnnotationCall<Identifier = String> {
 	/// Identifier of the constraint predicate
 	pub id: Identifier,
@@ -183,8 +188,9 @@ impl<Identifier: Display> Display for AnnotationCall<Identifier> {
 }
 
 ///Literal values as arguments to [`AnnotationCall`]
-#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
-#[serde(untagged)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(untagged))]
+#[derive(Clone, PartialEq, Debug)]
 pub enum AnnotationLiteral<Identifier = String> {
 	/// Basic FlatZinc literal (including annotation identifiers).
 	BaseLiteral(Literal<Identifier>),
@@ -202,8 +208,9 @@ impl<Idenfier: Display> Display for AnnotationLiteral<Idenfier> {
 }
 
 /// The argument type associated with [`Constraint`]
-#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
-#[serde(untagged)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(untagged))]
+#[derive(Clone, PartialEq, Debug)]
 pub enum Argument<Identifier = String> {
 	/// Sequence of [`Literal`]s
 	Array(Vec<Literal<Identifier>>),
@@ -238,20 +245,30 @@ impl<Identifier: Display> Display for Argument<Identifier> {
 /// information, in the form of [`Annotation`]s, from the MiniZinc model is
 /// stored in [`Array::ann`] when present. When [`Array::defined`] is set to
 /// `true`, then
-#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
-#[serde(rename = "array")]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(rename = "array"))]
+#[derive(Clone, PartialEq, Debug)]
 pub struct Array<Identifier = String> {
 	/// The values stored within the array literal
-	#[serde(rename = "a")]
+	#[cfg_attr(feature = "serde", serde(rename = "a"))]
 	pub contents: Vec<Literal<Identifier>>,
-	#[serde(default, skip_serializing_if = "Vec::is_empty")]
+	#[cfg_attr(
+		feature = "serde",
+		serde(default, skip_serializing_if = "Vec::is_empty")
+	)]
 	/// List of annotations
 	pub ann: Vec<Annotation<Identifier>>,
-	#[serde(default, skip_serializing_if = "is_false")]
+	#[cfg_attr(
+		feature = "serde",
+		serde(default, skip_serializing_if = "serde::is_false")
+	)]
 	/// This field is set to `true` when there is a constraint that has been marked as
 	/// defining this array.
 	pub defined: bool,
-	#[serde(default, skip_serializing_if = "is_false")]
+	#[cfg_attr(
+		feature = "serde",
+		serde(default, skip_serializing_if = "serde::is_false")
+	)]
 	/// This field is set to `true` when the array has been introduced by the
 	/// MiniZinc compiler, rather than being explicitly defined at the top-level
 	/// of the MiniZinc model.
@@ -279,18 +296,25 @@ impl<Identifier: Ord> Array<Identifier> {
 }
 
 /// An object depicting a constraint
-#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
-#[serde(rename = "constraint")]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(rename = "constraint"))]
+#[derive(Clone, PartialEq, Debug)]
 pub struct Constraint<Identifier = String> {
 	/// Identifier of the constraint predicate
 	pub id: Identifier,
 	/// Arguments of the constraint
 	pub args: Vec<Argument<Identifier>>,
 	/// Identifier of the variable that the constraint defines
-	#[serde(default, skip_serializing_if = "Option::is_none")]
+	#[cfg_attr(
+		feature = "serde",
+		serde(default, skip_serializing_if = "Option::is_none")
+	)]
 	pub defines: Option<Identifier>,
 	/// List of annotations
-	#[serde(default = "Vec::new", skip_serializing_if = "Vec::is_empty")]
+	#[cfg_attr(
+		feature = "serde",
+		serde(default = "Vec::new", skip_serializing_if = "Vec::is_empty")
+	)]
 	pub ann: Vec<Annotation<Identifier>>,
 }
 
@@ -321,44 +345,54 @@ impl<Identifier: Display> Display for Constraint<Identifier> {
 /// FlatZinc is (generally) a format produced by the MiniZinc compiler as a
 /// result of instantiating the parameter variables of a MiniZinc model and
 /// generating a solver-specific equisatisfiable model.
-#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(Clone, PartialEq, Debug)]
 pub struct FlatZinc<
 	Identifier = String,
 	VarMap = BTreeMap<Identifier, Variable<Identifier>>,
 	ArrayMap = BTreeMap<Identifier, Array<Identifier>>,
 > {
 	/// A mapping from decision variable `Identifier` to their definitions
-	#[serde(
-		default,
-		bound(
-			serialize = "Identifier: Serialize, for<'a> &'a VarMap: IntoIterator<Item = (&'a Identifier, &'a Variable<Identifier>)>",
-			deserialize = "Identifier: Deserialize<'de>, VarMap: FromIterator<(Identifier, Variable<Identifier>)>"
-		),
-		deserialize_with = "deserialize_key_value_object",
-		serialize_with = "serialize_key_value_object"
+	#[cfg_attr(
+		feature = "serde",
+		serde(
+			default,
+			bound(
+				serialize = "Identifier: Serialize, for<'a> &'a VarMap: IntoIterator<Item = (&'a Identifier, &'a Variable<Identifier>)>",
+				deserialize = "Identifier: Deserialize<'de>, VarMap: FromIterator<(Identifier, Variable<Identifier>)>"
+			),
+			deserialize_with = "serde::deserialize_key_value_object",
+			serialize_with = "serde::serialize_key_value_object"
+		)
 	)]
 	pub variables: VarMap,
 	/// A mapping from array `Identifier` to their definitions
-	#[serde(
-		default,
-		bound(
-			serialize = "Identifier: Serialize, for<'a> &'a ArrayMap: IntoIterator<Item = (&'a Identifier, &'a Array<Identifier>)>",
-			deserialize = "Identifier: Deserialize<'de>, ArrayMap: FromIterator<(Identifier, Array<Identifier>)>"
-		),
-		deserialize_with = "deserialize_key_value_object",
-		serialize_with = "serialize_key_value_object"
+	#[cfg_attr(
+		feature = "serde",
+		serde(
+			default,
+			bound(
+				serialize = "Identifier: Serialize, for<'a> &'a ArrayMap: IntoIterator<Item = (&'a Identifier, &'a Array<Identifier>)>",
+				deserialize = "Identifier: Deserialize<'de>, ArrayMap: FromIterator<(Identifier, Array<Identifier>)>"
+			),
+			deserialize_with = "serde::deserialize_key_value_object",
+			serialize_with = "serde::serialize_key_value_object"
+		)
 	)]
 	pub arrays: ArrayMap,
 	/// A list of (solver-specific) constraints, that must be satisfied in a solution.
-	#[serde(default)]
+	#[cfg_attr(feature = "serde", serde(default))]
 	pub constraints: Vec<Constraint<Identifier>>,
 	/// A list of all identifiers for which the solver must produce output for each solution
-	#[serde(default)]
+	#[cfg_attr(feature = "serde", serde(default))]
 	pub output: Vec<Identifier>,
 	/// A specification of the goal of solving the FlatZinc instance.
 	pub solve: SolveObjective<Identifier>,
 	/// The version of the FlatZinc serialization specification used
-	#[serde(default, skip_serializing_if = "String::is_empty")]
+	#[cfg_attr(
+		feature = "serde",
+		serde(default, skip_serializing_if = "String::is_empty")
+	)]
 	pub version: String,
 }
 
@@ -446,8 +480,9 @@ impl<Identifier: Ord + Display> Display for FlatZinc<Identifier> {
 // pub type Identifier = String;
 
 /// Literal values
-#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
-#[serde(untagged)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(untagged))]
+#[derive(Clone, PartialEq, Debug)]
 pub enum Literal<Identifier = String> {
 	/// Integer value
 	Int(i64),
@@ -457,22 +492,31 @@ pub enum Literal<Identifier = String> {
 	Identifier(Identifier),
 	/// Boolean value
 	Bool(bool),
-	#[serde(
-		serialize_with = "serialize_encapsulate_set",
-		deserialize_with = "deserialize_encapsulated_set"
+	#[cfg_attr(
+		feature = "serde",
+		serde(
+			serialize_with = "serde::serialize_encapsulate_set",
+			deserialize_with = "serde::deserialize_encapsulated_set"
+		)
 	)]
 	/// Set of integers, represented as a list of integer ranges
 	IntSet(RangeList<i64>),
-	#[serde(
-		serialize_with = "serialize_encapsulate_set",
-		deserialize_with = "deserialize_encapsulated_set"
+	#[cfg_attr(
+		feature = "serde",
+		serde(
+			serialize_with = "serde::serialize_encapsulate_set",
+			deserialize_with = "serde::deserialize_encapsulated_set"
+		)
 	)]
 	/// Set of floating point values, represented as a list of floating point
 	/// ranges
 	FloatSet(RangeList<f64>),
-	#[serde(
-		serialize_with = "serialize_encapsulate_string",
-		deserialize_with = "deserialize_encapsulated_string"
+	#[cfg_attr(
+		feature = "serde",
+		serde(
+			serialize_with = "serde::serialize_encapsulate_string",
+			deserialize_with = "serde::deserialize_encapsulated_string"
+		)
 	)]
 	/// String value
 	String(String),
@@ -493,18 +537,19 @@ impl<Identifier: Display> Display for Literal<Identifier> {
 }
 
 /// Goal of solving a FlatZinc instance
-#[derive(Default, Clone, PartialEq, Debug, Deserialize, Serialize)]
-#[serde(rename = "method")]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(rename = "method"))]
+#[derive(Default, Clone, PartialEq, Debug)]
 pub enum Method {
 	/// Find any solution
-	#[serde(rename = "satisfy")]
+	#[cfg_attr(feature = "serde", serde(rename = "satisfy"))]
 	#[default]
 	Satisfy,
 	/// Find the solution with the lowest objective value
-	#[serde(rename = "minimize")]
+	#[cfg_attr(feature = "serde", serde(rename = "minimize"))]
 	Minimize,
 	/// Find the solution with the highest objective value
-	#[serde(rename = "maximize")]
+	#[cfg_attr(feature = "serde", serde(rename = "maximize"))]
 	Maximize,
 }
 
@@ -519,18 +564,22 @@ impl Display for Method {
 }
 
 /// A specification of objective of a FlatZinc instance
-#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(Clone, PartialEq, Debug)]
 pub struct SolveObjective<Identifier = String> {
 	/// The type of goal
 	pub method: Method,
 	/// The variable to optimize, or `None` if [`SolveObjective::method`] is [`Method::Satisfy`]
-	#[serde(skip_serializing_if = "Option::is_none")]
+	#[cfg_attr(feature = "serde", serde(skip_serializing_if = "Option::is_none"))]
 	pub objective: Option<Literal<Identifier>>,
 	/// A list of annotations from the solve statement in the MiniZinc model
 	///
 	/// Note that this includes the search annotations if they are present in the
 	/// model.
-	#[serde(default, skip_serializing_if = "Vec::is_empty")]
+	#[cfg_attr(
+		feature = "serde",
+		serde(default, skip_serializing_if = "Vec::is_empty")
+	)]
 	pub ann: Vec<Annotation<Identifier>>,
 }
 
@@ -617,308 +666,4 @@ pub struct Variable<Identifier = String> {
 	/// MiniZinc compiler, rather than being explicitly defined at the top-level
 	/// of the MiniZinc model.
 	pub introduced: bool,
-}
-
-impl<Identifier: Serialize> Serialize for Variable<Identifier> {
-	fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-		#[derive(Serialize)]
-		#[serde(rename = "variable")]
-		struct VariableRepr<'a, Identifier> {
-			#[serde(rename = "type")]
-			ty: BaseType,
-			#[serde(skip_serializing_if = "Option::is_none")]
-			domain: Option<VariableDomain>,
-			#[serde(rename = "rhs", skip_serializing_if = "Option::is_none")]
-			value: Option<&'a Literal<Identifier>>,
-			#[serde(default, skip_serializing_if = "Vec::is_empty")]
-			ann: &'a Vec<Annotation<Identifier>>,
-			#[serde(default, skip_serializing_if = "is_false")]
-			defined: bool,
-			#[serde(default, skip_serializing_if = "is_false")]
-			introduced: bool,
-		}
-
-		let (ty, domain) = match &self.ty {
-			Type::Bool => (BaseType::Bool, None),
-			Type::Int(domain) => (BaseType::Int, domain.clone().map(VariableDomain::Int)),
-			Type::Float(domain) => (BaseType::Float, domain.clone().map(VariableDomain::Float)),
-			Type::IntSet(domain) => (BaseType::IntSet, domain.clone().map(VariableDomain::Int)),
-		};
-
-		VariableRepr {
-			ty,
-			domain,
-			value: self.value.as_ref(),
-			ann: &self.ann,
-			defined: self.defined,
-			introduced: self.introduced,
-		}
-		.serialize(serializer)
-	}
-}
-
-impl<'de, Identifier: Deserialize<'de>> Deserialize<'de> for Variable<Identifier> {
-	fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-		#[derive(Deserialize)]
-		#[serde(rename = "variable")]
-		#[serde(bound(deserialize = "Identifier: Deserialize<'de>"))]
-		struct VariableRepr<Identifier> {
-			#[serde(rename = "type")]
-			ty: BaseType,
-			#[serde(skip_serializing_if = "Option::is_none")]
-			domain: Option<VariableDomain>,
-			#[serde(rename = "rhs", skip_serializing_if = "Option::is_none")]
-			value: Option<Literal<Identifier>>,
-			#[serde(default, skip_serializing_if = "Vec::is_empty")]
-			ann: Vec<Annotation<Identifier>>,
-			#[serde(default, skip_serializing_if = "is_false")]
-			defined: bool,
-			#[serde(default, skip_serializing_if = "is_false")]
-			introduced: bool,
-		}
-
-		let repr = VariableRepr::deserialize(deserializer)?;
-		let ty = match (repr.ty, repr.domain) {
-			(BaseType::Bool, None) => Type::Bool,
-			(BaseType::Bool, Some(_)) => {
-				return Err(SerdeError::custom("bool variables cannot have a domain"));
-			}
-			(BaseType::Int, None) => Type::Int(None),
-			(BaseType::Int, Some(VariableDomain::Int(domain))) => Type::Int(Some(domain)),
-			(BaseType::Int, Some(VariableDomain::Float(_))) => {
-				return Err(SerdeError::custom("int variables require an int domain"));
-			}
-			(BaseType::Float, None) => Type::Float(None),
-			(BaseType::Float, Some(VariableDomain::Float(domain))) => Type::Float(Some(domain)),
-			(BaseType::Float, Some(VariableDomain::Int(_))) => {
-				return Err(SerdeError::custom("float variables require a float domain"));
-			}
-			(BaseType::IntSet, None) => Type::IntSet(None),
-			(BaseType::IntSet, Some(VariableDomain::Int(domain))) => Type::IntSet(Some(domain)),
-			(BaseType::IntSet, Some(VariableDomain::Float(_))) => {
-				return Err(SerdeError::custom(
-					"set of int variables require an int domain",
-				));
-			}
-		};
-
-		Ok(Variable {
-			ty,
-			value: repr.value,
-			ann: repr.ann,
-			defined: repr.defined,
-			introduced: repr.introduced,
-		})
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use std::{
-		collections::{BTreeMap, HashMap},
-		fs::File,
-		io::{BufReader, Read},
-		path::Path,
-	};
-
-	use expect_test::ExpectFile;
-	use ustr::Ustr;
-
-	use crate::{
-		Annotation, AnnotationArgument, AnnotationCall, AnnotationLiteral, Array, FlatZinc,
-		Literal, Method, RangeList, SolveObjective, Type, Variable,
-	};
-
-	test_file!(documentation_example);
-	test_file!(encapsulated_string);
-	test_file!(float_sets);
-	test_file!(set_literals);
-	test_file!(unit_test_example);
-
-	fn test_successful_serialization(file: &Path, exp: ExpectFile) {
-		let rdr = BufReader::new(File::open(file).unwrap());
-		let fzn: FlatZinc = serde_json::from_reader(rdr).unwrap();
-		exp.assert_debug_eq(&fzn);
-		let fzn2: FlatZinc = serde_json::from_str(&serde_json::to_string(&fzn).unwrap()).unwrap();
-		assert_eq!(fzn, fzn2)
-	}
-
-	macro_rules! test_file {
-		($file: ident) => {
-			#[test]
-			fn $file() {
-				test_successful_serialization(
-					std::path::Path::new(&format!("./corpus/json/{}.fzn.json", stringify!($file))),
-					expect_test::expect_file![&format!(
-						"../corpus/json/{}.debug.txt",
-						stringify!($file)
-					)],
-				)
-			}
-		};
-	}
-	pub(crate) use test_file;
-
-	#[test]
-	fn test_ident_no_copy() {
-		let mut rdr = BufReader::new(
-			File::open(Path::new("./corpus/json/documentation_example.fzn.json")).unwrap(),
-		);
-		let mut content = String::new();
-		let _ = rdr.read_to_string(&mut content).unwrap();
-
-		let fzn: FlatZinc<&str> = serde_json::from_str(&content).unwrap();
-		expect_test::expect_file!["../corpus/json/documentation_example.debug.txt"]
-			.assert_debug_eq(&fzn)
-	}
-
-	#[test]
-	fn test_ident_interned() {
-		let rdr = BufReader::new(
-			File::open(Path::new("./corpus/json/documentation_example.fzn.json")).unwrap(),
-		);
-		let fzn: FlatZinc<Ustr> = serde_json::from_reader(rdr).unwrap();
-		expect_test::expect_file!["../corpus/json/documentation_example.debug_ustr.txt"]
-			.assert_debug_eq(&fzn)
-	}
-
-	#[test]
-	fn test_hashmap_backed_maps_deserialize() {
-		type HashMapFlatZinc =
-			FlatZinc<String, HashMap<String, Variable<String>>, HashMap<String, Array<String>>>;
-
-		let mut rdr = BufReader::new(
-			File::open(Path::new("./corpus/json/documentation_example.fzn.json")).unwrap(),
-		);
-		let mut content = String::new();
-		let _ = rdr.read_to_string(&mut content).unwrap();
-
-		let fzn: HashMapFlatZinc = serde_json::from_str(&content).unwrap();
-
-		let fzn2: HashMapFlatZinc = {
-			let json = serde_json::to_string(&fzn).unwrap();
-			serde_json::from_str(&json).unwrap()
-		};
-		assert_eq!(fzn, fzn2);
-	}
-
-	#[test]
-	fn test_vec_backed_maps_deserialize_from_object_shape() {
-		type VecFlatZinc =
-			FlatZinc<String, Vec<(String, Variable<String>)>, Vec<(String, Array<String>)>>;
-
-		let mut rdr = BufReader::new(
-			File::open(Path::new("./corpus/json/documentation_example.fzn.json")).unwrap(),
-		);
-		let mut content = String::new();
-		let _ = rdr.read_to_string(&mut content).unwrap();
-
-		let fzn: VecFlatZinc = serde_json::from_str(&content).unwrap();
-		assert!(!fzn.variables.is_empty());
-		assert!(!fzn.arrays.is_empty());
-	}
-
-	#[test]
-	fn test_default_with_custom_map_types() {
-		let fzn = FlatZinc::<
-			String,
-			HashMap<String, Variable<String>>,
-			Vec<(String, Array<String>)>,
-		>::default();
-		assert!(fzn.variables.is_empty());
-		assert!(fzn.arrays.is_empty());
-		assert!(fzn.constraints.is_empty());
-		assert!(fzn.output.is_empty());
-		assert_eq!(fzn.version, "1.0");
-	}
-
-	#[test]
-	fn test_print_flatzinc() {
-		let mut rdr = BufReader::new(
-			File::open(Path::new("./corpus/json/documentation_example.fzn.json")).unwrap(),
-		);
-		let mut content = String::new();
-		let _ = rdr.read_to_string(&mut content).unwrap();
-
-		let fzn: FlatZinc<&str> = serde_json::from_str(&content).unwrap();
-		expect_test::expect_file!["../corpus/fzn/documentation_example.fzn"]
-			.assert_eq(&fzn.to_string());
-
-		let ann: Annotation<&str> = Annotation::Call(AnnotationCall {
-			id: "bool_search",
-			args: vec![
-				AnnotationArgument::Literal(AnnotationLiteral::BaseLiteral(Literal::Identifier(
-					"input_order",
-				))),
-				AnnotationArgument::Literal(AnnotationLiteral::BaseLiteral(Literal::Identifier(
-					"indomain_min",
-				))),
-			],
-		});
-		assert_eq!(ann.to_string(), "::bool_search(input_order, indomain_min)");
-
-		let ty = Type::Bool;
-		assert_eq!(ty.to_string(), "bool");
-		let ty = Type::Int(None);
-		assert_eq!(ty.to_string(), "int");
-		let ty = Type::Float(None);
-		assert_eq!(ty.to_string(), "float");
-		let ty = Type::IntSet(None);
-		assert_eq!(ty.to_string(), "set of int");
-		let ty = Type::Float(Some(RangeList::from(1.0..=4.0)));
-		assert_eq!(ty.to_string(), "1.0..4.0");
-
-		let lit = Literal::<&str>::Int(1);
-		assert_eq!(lit.to_string(), "1");
-		let lit = Literal::<&str>::Float(1.0);
-		assert_eq!(lit.to_string(), "1.0");
-		let lit = Literal::<&str>::Identifier("x");
-		assert_eq!(lit.to_string(), "x");
-		let lit = Literal::<&str>::Bool(true);
-		assert_eq!(lit.to_string(), "true");
-		let lit = Literal::<&str>::IntSet(RangeList::from(2..=3));
-		assert_eq!(lit.to_string(), "2..3");
-		let lit = Literal::<&str>::FloatSet(RangeList::from(2.0..=3.0));
-		assert_eq!(lit.to_string(), "2.0..3.0");
-		let lit = Literal::<&str>::String(String::from("hello"));
-		assert_eq!(lit.to_string(), "\"hello\"");
-
-		let fzn = FlatZinc {
-			variables: BTreeMap::from([(
-				"x",
-				Variable {
-					ty: Type::IntSet(None),
-					ann: vec![Annotation::Atom("special")],
-					defined: false,
-					introduced: true,
-					value: Some(Literal::IntSet(RangeList::from(1..=4))),
-				},
-			)]),
-			arrays: BTreeMap::from([(
-				"y",
-				Array {
-					ann: vec![Annotation::Atom("special")],
-					contents: vec![Literal::Int(1), Literal::Int(2), Literal::Int(3)],
-					introduced: true,
-					defined: true,
-				},
-			)]),
-			output: vec!["y"],
-			..Default::default()
-		};
-		assert_eq!(
-			fzn.to_string(),
-			"var set of int: x ::var_is_introduced ::special = 1..4;\narray[1..3] of int: y ::output_array([1..3]) ::is_defined_var ::var_is_introduced ::special = [1, 2, 3];\nsolve satisfy;\n"
-		);
-
-		let sat = SolveObjective {
-			method: Method::Minimize,
-			ann: vec![ann],
-			objective: Some(Literal::Identifier("x")),
-		};
-		assert_eq!(
-			sat.to_string(),
-			"solve ::bool_search(input_order, indomain_min) minimize x"
-		);
-	}
 }
